requirements.txt
streamlit>=1.30.0
PyMuPDF>=1.23.0
langchain-google-genai>=1.0.0
langchain-core>=0.2.0
python-dotenv>=1.0.0
altair>=5.0.0
pandas>=2.0.0
